mcp_project_main/requirements.txt
streamlit
pandas
python-dotenv
openai
tavily-python
google-generativeai